Added catalogpath code to barrier_master.py Removed automatic rerun from circuitscape code Minor changes to s6 voltage maps added to s8, other minor changes

+++ toolbox/scripts/circuitscape_master.py
@@ -82,7 +82,11 @@ def circuitscape_master():
                 msg = ('ERROR: Resistance raster is required for pinch point'
                         ' analyses, but was not found.')
                 lu.raise_error(msg)
-
+            
+            desc = arcpy.Describe(cfg.RESRAST_IN)
+            if hasattr(desc, "catalogPath"):
+                cfg.RESRAST_IN = arcpy.Describe(cfg.RESRAST_IN).catalogPath
+            
             arcpy.env.extent = cfg.RESRAST_IN
             arcpy.env.snapRaster = cfg.RESRAST_IN
             gprint('\nMaking local copy of resistance raster.')
@@ -114,17 +118,19 @@ def circuitscape_master():
 
 #aaa            lu.clean_out_workspace(cfg.PINCHGDB)
 #aaa            lu.delete_data(cfg.PINCHGDB) 
-            for i in range (1,2):
-                try:
-                    s8.STEP8_calc_pinchpoints()            
-                    break
-                except:
-                    gprint('******************** Run failed ****************************')
-                    gprint('Trying one more time in 20 seconds.')
-                    lu.snooze(20)
-                    s8.STEP8_calc_pinchpoints()
-
-            lu.delete_dir(cfg.SCRATCHDIR)
+            # for i in range (1,2):
+                # try:
+                    # s8.STEP8_calc_pinchpoints()            
+                    # break
+                # except:
+                    # gprint('******************** Run failed ****************************')
+                    # gprint('Trying one more time in 20 seconds.')
+                    # lu.snooze(20)
+                    # s8.STEP8_calc_pinchpoints()
+            s8.STEP8_calc_pinchpoints()            
+            
+            if not cfg.SAVE_TEMP_FILES:
+                lu.delete_dir(cfg.SCRATCHDIR)
             if not cfg.SAVECIRCUITDIR:
                 lu.delete_dir(cfg.CIRCUITBASEDIR)
